+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - Updated readme for installation part ([#225])(https://github.com/chingu-x/chingu-dashboard-be/pull/225)
 
 ### Fixed
+- fixed POST voyages/teams/{teamId}/techs bug , verify that categoryId is owned by correct team ([#228](https://github.com/chingu-x/chingu-dashboard-be/pull/228))
 
 ### Removed